hexa-ai-logo-generator/functions/requirements.txt
firebase-admin 
functions-framework
firebase-functions
requests
Pillow